+++ pkg/client/userd/trafficmgr/intercept.go
@@ -278,8 +278,7 @@ func interceptError(tp rpc.InterceptError, err error) *rpc.InterceptResult {
 // only if the returned rpc.InterceptResult is nil. The returned runtime.Object is either nil, indicating a local
 // intercept, or the workload for the intercept.
 func (tm *TrafficManager) CanIntercept(c context.Context, ir *rpc.CreateInterceptRequest) (*rpc.InterceptResult, k8sapi.Workload, *ServiceProps) {
-	tm.WaitForNSSync(c)
-	tm.wlWatcher.waitForSync(c)
+	tm.waitForSync(c)
 	spec := ir.Spec
 	spec.Namespace = tm.ActualNamespace(spec.Namespace)
 	if spec.Namespace == "" {