+++ CHANGELOG.md
@@ -1,8 +1,24 @@
 # Changelog
 
-### 2.5.5 (TBD)
+### 2.5.6 (TBD)
 
 - Feature: The flag `--preview-url-add-request-headers` can be used to inject key value pair headers in every preview page request
+
+### 2.5.5 (TBD)
+
+- Change: The traffic-manager now requires permissions to read pods across namespaces even if installed with limited permissions
+
+- Bugfix: The DNS resolver used on Linux with systemd-resolved now flushes the cache when the search path changes.
+
+- Bugfix: The `telepresence list` command will produce a correct listing even when not preceded by a `telepresence connect`.
+
+- Bugfix: The root daemon will no longer get into a bad state when a disconnect is rapidly followed by a new connect.
+
+- Bugfix: The client will now only watch agents from accessible namespaces, and is also constrained to namespaces  explicitly mapped
+  using the `connect` command's `--mapped-namespaces` flag.
+
+- Bugfix: The `gather-logs` command will only gather traffic-agent logs from accessible namespaces, and is also constrained to namespaces
+  explicitly mapped using the `connect` command's `--mapped-namespaces` flag.
 
 ### 2.5.4 (March 29, 2022)